Refactor Song and Movie storage (local and cloud).
Simplify the APIs and drive changes all the way down the stack so that
we represent local and cloud movies/songs in a more intuitive fashion.
Along the way, use the file size as part of the hash so that we can
detect when a file has the same name, but has different data.

There's still lots of overlap between Movies and Songs classes.