Simplify the segue flow by not allowing Songs to go back to Welcome.
This will ultimately make all of the modal dismissals be much easier for now.
It also means, however, that we're getting rid of the ability for you to cue
up a welcome animation since there's no way to go back from Songs to Welcome.
I think that's an ok tradeoff for now since we rarely use it and the whole
segue and segue unwind logic is tedious and confusing.